now saving spritekit location

+++ Contained-iOS6/Contained-iOS6/Settings.swift
@@ -13,6 +13,6 @@ class Settings {
 	var shouldRoll = false
 	var shouldZoom = false
 	
-	var position_x: CGFloat = CGFloat(0)
-	var position_y: CGFloat = CGFloat(0)
+	var position_x: CGFloat = 0
+	var position_y: CGFloat = 0
 }
